scan-query: Use long as limit. (#4081)
* scan-query: Use long instead of int as limit type

* Use MAX_INSTANT queryTimeout, if timeout == 0

+++ extensions-contrib/scan-query/src/main/java/io/druid/query/scan/ScanQueryQueryToolChest.java
@@ -50,7 +50,7 @@ public class ScanQueryQueryToolChest extends QueryToolChest<ScanResultValue, Sca
       )
       {
         ScanQuery scanQuery = (ScanQuery) query;
-        if (scanQuery.getLimit() == Integer.MAX_VALUE) {
+        if (scanQuery.getLimit() == Long.MAX_VALUE) {
           return runner.run(query, responseContext);
         }
         return new BaseSequence<>(
